refactor: InputService -> InputController로 변경
- Service가 아닌 Controller의 역할이 맞다고 판단
