+++ src/controllers/LottoController.js
@@ -1,4 +1,4 @@
-import InputService from "../services/InputService.js";
+import InputController from "./InputController.js";
 import issueLottos from "../domains/issueLottos.js";
 import OutputView from "../views/OutputView.js";
 import WinningStatistics from "../domains/WinningStatistics.js";
@@ -7,17 +7,18 @@ import { YES } from "../constants/constants.js";
 class Controller {
   async start() {
     await this.#runLottoGame();
-    const restartConfirm = await InputService.getValidRestartConfirm();
+    const restartConfirm = await InputController.getValidRestartConfirm();
     if (restartConfirm === YES) await this.start();
   }
 
   async #runLottoGame() {
-    const purchaseAmount = await InputService.getValidPurchaseAmount();
+    const purchaseAmount = await InputController.getValidPurchaseAmount();
     const lottos = issueLottos(purchaseAmount);
     OutputView.printIssuedLottos(lottos);
 
-    const winningNumbers = await InputService.getValidWinningNumbers();
-    const bonusNumber = await InputService.getValidBonusNumber(winningNumbers);
+    const winningNumbers = await InputController.getValidWinningNumbers();
+    const bonusNumber =
+      await InputController.getValidBonusNumber(winningNumbers);
 
     const winningStatistics = new WinningStatistics(lottos);
     winningStatistics.calculateWinningResults(winningNumbers, bonusNumber);
